+++ packages/wallet-react-ui/src/signing/components/Section/MessageDetails/index.tsx
@@ -1,4 +1,5 @@
-import { DataRow } from '../../DataRow'
+import { DataRow } from '@zerodev/react-ui'
+import { camelCaseToTitle } from '../../../../shared/utils/common'
 import { Section } from '..'
 
 export interface MessageDetailsProps {
@@ -9,7 +10,9 @@ export function MessageDetails({ details }: MessageDetailsProps) {
   return (
     <Section title="Message Details" iconName="message">
       {Object.entries(details).map(([label, value]) => (
-        <DataRow key={label} label={label} value={value} />
+        // Object keys are raw camelCase (e.g. "fromAddress"); the primitive
+        // DataRow renders labels verbatim, so title-case them here.
+        <DataRow key={label} label={camelCaseToTitle(label)} value={value} />
       ))}
     </Section>
   )
