+++ packages/wallet-react-ui/src/signing/pages/Erc20Approval.tsx
@@ -1,9 +1,8 @@
-import { Text } from '@zerodev/react-ui'
+import { DataRow, DataRowSkeleton, Icon, Text } from '@zerodev/react-ui'
 import { type Address, erc20Abi, formatUnits, type Hex, maxUint256 } from 'viem'
 import { useReadContract } from 'wagmi'
 import { shortenHex } from '../../shared/utils/common'
 import { ArrowCardPair } from '../components/ArrowCardPair'
-import { DataRow, DataRowSkeleton } from '../components/DataRow'
 import { InfoCard } from '../components/InfoCard'
 import { Section } from '../components/Section'
 import { SigningLayout } from '../components/SigningLayout'
@@ -112,7 +111,12 @@ export function Erc20Approval({
                 <DataRow
                   label="Fee"
                   value={formatGasFee(gasEstimate)}
-                  iconName="gasStation"
+                  trailing={
+                    <Icon
+                      name="gasStation"
+                      className="zd:w-4 zd:h-4 zd:text-solarOrange"
+                    />
+                  }
                 />
               ) : (
                 <DataRowSkeleton label="Fee" />
