+++ packages/wallet-react-ui/src/signing/pages/GenericRequest.tsx
@@ -1,4 +1,4 @@
-import { Text } from '@zerodev/react-ui'
+import { DataRow, DataRowSkeleton, Icon, Text } from '@zerodev/react-ui'
 import {
   type Address,
   formatEther,
@@ -9,7 +9,6 @@ import {
 } from 'viem'
 import { shortenHex } from '../../shared/utils/common'
 import type { Request } from '../../types.js'
-import { DataRow, DataRowSkeleton } from '../components/DataRow'
 import { Section } from '../components/Section'
 import { SigningLayout } from '../components/SigningLayout'
 import { useGasEstimate } from '../hooks/useGasEstimate'
@@ -112,7 +111,12 @@ function GenericSendTransaction({
               <DataRow
                 label="Fee"
                 value={formatGasFee(gasEstimate)}
-                iconName="gasStation"
+                trailing={
+                  <Icon
+                    name="gasStation"
+                    className="zd:w-4 zd:h-4 zd:text-solarOrange"
+                  />
+                }
               />
             ) : (
               <DataRowSkeleton label="Fee" />
