+++ packages/wallet-react-ui/src/signing/pages/MintNft.tsx
@@ -1,7 +1,6 @@
-import { Text } from '@zerodev/react-ui'
+import { DataRow, DataRowSkeleton, Icon, Text } from '@zerodev/react-ui'
 import type { Address, Hex } from 'viem'
 import { useReadContract } from 'wagmi'
-import { DataRow, DataRowSkeleton } from '../components/DataRow'
 import { InfoCard } from '../components/InfoCard'
 import { Section } from '../components/Section'
 import { SigningLayout } from '../components/SigningLayout'
@@ -83,7 +82,12 @@ export function MintNft({ contract, data, confirm, reject }: MintNftProps) {
                 <DataRow
                   label="Fee"
                   value={formatGasFee(gasEstimate)}
-                  iconName="gasStation"
+                  trailing={
+                    <Icon
+                      name="gasStation"
+                      className="zd:w-4 zd:h-4 zd:text-solarOrange"
+                    />
+                  }
                 />
               ) : (
                 <DataRowSkeleton label="Fee" />
