+++ packages/wallet-react-ui/src/signing/pages/SignTypedData.tsx
@@ -1,7 +1,6 @@
-import { Text } from '@zerodev/react-ui'
+import { DataRow, Text } from '@zerodev/react-ui'
 import type { Hex } from 'viem'
 import { shortenHex } from '../../shared/utils/common'
-import { DataRow } from '../components/DataRow'
 import { Section } from '../components/Section'
 import { SigningLayout } from '../components/SigningLayout'
 import { TypedDataMessage } from '../components/TypedDataMessage'
